Adding new method to class spider
Adding function to check executable. It will print if the executable exists and raise an error if it doesn't. It will add the name of the executable if someone give just the bin folder where the executable is. If just the name of the executable is given for the path, running which to see the output if the file exists.