Create XnatShareSession (#424)
* Create XnatShareSession

NOT DONE

Right now I'm just getting this to actually work with a test, then I'll go in and make it like the other tools.

* Update XnatShareSession

* Update XnatShareSession

This almost works. It shares into the VUIIS_ABCD project, BUT it shares the ACCESSION NUMBER rather than the subject/experiment.

For instance, VUIISXNAT04_S17846 instead of 212533 for the subject and VUIISXNAT04_E232899 instead of 212533 for the experiment.

When I go into the subject, I also do not see any scans or experiments listed, but when I click delete, they are listed as something that can be deleted.

* Update XnatShareSession

* Update XnatShareSession

* Update XnatShareSession

* Update XnatShareSession

* Update XnatShareSession

* Update XnatShareSession

* Update XnatShareSession

Adding imports. I'm not sure I updated this with the final version, so just doing it now.

* Update XnatShareSession