Create XnatDownloadGUI
Not sure this is entirely working yet. Also doesn't have all the options. I'm sure this can be approved upon readability wise as well. 

It's to a point where it puts a file in the directory given though. 

One problem I know of currently is that if you give it scan parameters, it will still look at ALL assessor info (and vice versa; assessor info, still looks at ALL scan info). Should be an easy fix, but just getting this somewhere.

# Parameters :                                                 #
#     Directory            -> /Users/wilduett/Downloads        #
#     Project(s)           -> VUIIS_ABCD                       #
#     Subject(s)           -> 001                              #
#     Session(s)           -> 001_3TA_1                        #
#     Scan types           -> 201                              #
#     Resources Scan       -> NIFTI                            #
#     Process types        -> all                              #
#     Resources Process    -> all                              #